# openapi/erp-api.yaml
openapi: 3.1.0
info:
  title: ERP Platform API
  description: >-
    Modular monolith ERP. Operations marked x-implementation-status: roadmap are contracts only and are NOT implemented.
    Public routes: health, status, register and login. Refresh/logout require a refresh credential. Login requires x-tenant-id. Protected routes require JWT plus
    matching tenant context and an active membership. No default tenant. Responses use data/meta/correlationId or
    error/correlationId, without a success flag. Company/branch headers are contextual metadata, not independently
    authorized scopes. Business company routes also require an active persisted company and explicit company membership.
  version: v1
  contact:
    name: ERP Platform Team
    email: dev@erp.local
    url: https://erp.local
  license:
    name: Proprietary
    url: https://erp.local/license
servers:
  - url: http://localhost:3000
    description: Local development; paths include /api/v1 where applicable
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        HS256 access JWT with sub, sid, tenantId, roles, iat and exp. Session, user and membership are checked in
        MongoDB on every request. Revocation denies subsequent checks immediately; in-flight work is not rolled back.
    refreshCookie:
      type: apiKey
      in: cookie
      name: erp_refresh
      description: HttpOnly SameSite Strict cookie scoped to /api/v1/auth; Secure in production. Web writes require an exact trusted Origin and x-session-client web.
  schemas:
    Error:
      type: object
      required:
        - error
        - correlationId
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        correlationId:
          type: string
      example:
        error:
          code: FORBIDDEN
          message: Forbidden
        correlationId: example-request-id
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_SERVER_ERROR
            - BAD_REQUEST
            - TENANT_ACCESS_DENIED
            - INSUFFICIENT_STOCK
            - INVENTORY_INSUFFICIENT_STOCK
            - UNAUTHORIZED_TRANSITION
            - FISCAL_PERIOD_CLOSED
            - DUPLICATE_RESOURCE
            - RESOURCE_NOT_FOUND
            - VALIDATION_SCHEMA_ERROR
            - IDEMPOTENCY_KEY_EXISTS
            - RATE_LIMIT_EXCEEDED
            - EXTERNAL_SERVICE_ERROR
            - TIMEOUT_ERROR
            - FEATURE_DISABLED
            - PERMISSION_DENIED
            - TENANT_CONTEXT_MISSING
            - INTERNAL_ERROR
            - RATE_LIMITED
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
...
    CustomerCreate:
      type: "object"
      additionalProperties: false
      required: ["type","name"]
      properties:
        type:
          type: "string"
          enum: ["company","individual"]
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        email:
          type: "string"
          maxLength: 254
          format: "email"
        phone:
          type: "string"
          maxLength: 50
        taxId:
          type: "string"
          maxLength: 100
...
    Customer:
      type: "object"
      additionalProperties: false
      required: ["customerId","tenantId","companyId","name","type","status","version","createdAt","updatedAt"]
      properties:
        type:
          type: "string"
          enum: ["company","individual"]
        name:
          type: "string"
          maxLength: 200
          minLength: 1
        email:
          type: "string"
          maxLength: 254
          format: "email"
        phone:
          type: "string"
          maxLength: 50
        taxId:
          type: "string"
          maxLength: 100
        customerId:
          type: "string"
          format: "uuid"
        tenantId:
          type: "string"
          maxLength: 200
        companyId:
          type: "string"
          format: "uuid"
        status:
          type: "string"
          enum: ["active","inactive","blocked"]
        version:
          type: "integer"
          minimum: 1
          maximum: 9007199254740990
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
        createdBy:
          type: "string"
          maxLength: 200
        updatedBy:
          type: "string"
          maxLength: 200
...
    CustomerPage:
      type: "object"
      additionalProperties: false
      required: ["items","total","page","limit","totalPages"]
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: "integer"
          minimum: 0
        page:
          type: "integer"
          minimum: 1
        limit:
          type: "integer"
          minimum: 1
          maximum: 100
        totalPages:
          type: "integer"
          minimum: 0
...
  parameters:
...
    CompanyIdPathParam:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The company ID
...
    TenantIdHeader:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
      description: >-
        Explicit tenant selector. Must equal the JWT tenant on protected routes; missing returns 401
        TENANT_CONTEXT_MISSING, mismatch returns 403.
...
paths:
...
  "/api/v1/companies/{companyId}/crm/customers":
    get:
      operationId: "listCustomers"
      summary: "Search company customers"
      description: "Requires crm.customer.read. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. Literal case-insensitive search over name/email. Names/emails are not unique. Count and page are separate reads."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"},{"name":"page","in":"query","schema":{"type":"integer","minimum":1,"maximum":1000000,"default":1}},{"name":"limit","in":"query","schema":{"type":"integer","minimum":1,"maximum":100,"default":20}},{"name":"search","in":"query","schema":{"type":"string","maxLength":200}},{"name":"status","in":"query","schema":{"type":"string","enum":["active","inactive","blocked"]}}]
      responses:
        "200":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/CustomerPage"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: "createCustomer"
      summary: "Register customer"
      description: "Requires crm.customer.write. Live session and tenant membership required. Business company paths require an active company and explicit company membership; headers never grant access. Initial status active and version 1. Mutation and audit are one transaction."
      x-implementation-status: "implemented"
      parameters: [{"$ref":"#/components/parameters/TenantIdHeader"},{"$ref":"#/components/parameters/CompanyIdPathParam"}]
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/CustomerCreate"
      responses:
        "201":
          description: "Successful operation"
          content:
            "application/json":
              schema:
                type: "object"
                required: ["data","correlationId"]
                properties:
                  data:
                    $ref: "#/components/schemas/Customer"
                  correlationId:
                    type: "string"
        "400":
          description: "Invalid or protected input"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "Missing or expired session"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "Permission, tenant or company access denied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "Scoped resource not found"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "Version, duplicate key or idempotency conflict"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: "Database or transaction failure; no success implied"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Error"
...